Components

+++ src/components.ur.robotics/src/AXOpen.Components.Ur.Robotics/AxoUrCb3_v_3_x_x/AxoUrCb3_v_3_x_x.cs
@@ -219,14 +219,14 @@ namespace AXOpen.Components.Ur.Robotics
     }
     public partial class AxoUrRobotics_Component_Status_v_1_x_x : AXOpen.Components.Robotics.AxoRobot_Status
     {
-        Dictionary<uint, string> errorDescriptionDict = new Dictionary<uint, string>();
-        Dictionary<uint, string> actionDescriptionDict = new Dictionary<uint, string>();
+        Dictionary<ulong, string> errorDescriptionDict = new Dictionary<ulong, string>();
+        Dictionary<ulong, string> actionDescriptionDict = new Dictionary<ulong, string>();
 
         public string ErrorDescription
         {
             get
             {
-                if (errorDescriptionDict == null) { errorDescriptionDict = new Dictionary<uint, string>(); }
+                if (errorDescriptionDict == null) { errorDescriptionDict = new Dictionary<ulong, string>(); }
                 if (errorDescriptionDict.Count == 0)
                 {
                     errorDescriptionDict.Add(0, "   ");
@@ -382,7 +382,7 @@ namespace AXOpen.Components.Ur.Robotics
         {
             get
             {
-                if (actionDescriptionDict == null) { actionDescriptionDict = new Dictionary<uint, string>(); }
+                if (actionDescriptionDict == null) { actionDescriptionDict = new Dictionary<ulong, string>(); }
                 if (actionDescriptionDict.Count == 0)
                 {
                     actionDescriptionDict.Add(0, "   ");
